Updates to the main menu and the kill-feed (see 'extended info')
[Main menu]
Moved the console button to the left (next to the "Store" button); it also adopts its design. Other changes include fixing up the navigation-through-buttons with the keyboard as the TF Team left obsolete keylinks in it, so some buttons were previously "redirecting" to non-existing buttons.

[Kill Feed]
Updated the RED/BLU color schemes and made them fully opaque.